JAMES-1877 Create mailDelivrer
Its responsibility :
 - Resolve MXs addresses for a set of domains
 - Try host to host delivery
 - Report execution
